refactor: Remove unused API endpoints and services
