--- src/api/endpoints/auth.ts
@@ -1,7 +1,0 @@
-const authEndpoints = {
-  signin: 'auth/signin',
-  logout: 'auth/logout',
-  refreshToken: 'auth/refresh-token',
-};
-
-export default authEndpoints;
