--- src/api/endpoints/index.ts
@@ -1,3 +1,0 @@
-export { default as roleEndpoints } from './roles';
-export { default as userEndpoints } from './users';
-export { default as authEndpoints } from './auth';
