--- src/api/endpoints/specialty.ts
@@ -1,5 +1,0 @@
-const specialtyEndpoints = {
-  getSpecialties: '/specialty',
-};
-
-export default specialtyEndpoints;
