--- src/api/endpoints/users.ts
@@ -1,6 +1,0 @@
-const userEndpoints = {
-  getAUsers: 'user',
-  getUser: (id: number) => `user/${id}`,
-};
-
-export default userEndpoints;
