--- src/api/services/index.ts
@@ -1,3 +1,0 @@
-export * as specialtyService from './specialty';
-export * as authService from './authService';
-export * as userService from './userService';
