+++ src/modules/shared/components/ExpiredSession/index.tsx
@@ -1,17 +1,16 @@
 import { Box, Button, Fade, Modal } from '@mui/material';
 
-import useAuth from '../../../../hooks/useAuth';
 import './styles.css';
 
 const ExpiredSession = () => {
-  const { showAuthenticatedModa, handleLogOut, handleRefreshToken } = useAuth();
+  // const { showAuthenticatedModa, handleLogOut, handleRefreshToken } = useAuth();
   return (
     <Modal
-      open={showAuthenticatedModa}
+      open={false}
       aria-labelledby="Modal para renovar la sesión"
       aria-describedby="Este modal es usado para cuando la sesión esta expirada y puede renovar"
     >
-      <Fade in={showAuthenticatedModa}>
+      <Fade in={false}>
         <Box component="div" className="expired-session">
           <Box component="div" className="content">
             <h1>⚠️ La sesión ha expirado</h1>
@@ -21,14 +20,10 @@ const ExpiredSession = () => {
             </p>
 
             <Box className="actions">
-              <Button
-                color="primary"
-                variant="contained"
-                onClick={handleRefreshToken}
-              >
+              <Button color="primary" variant="contained" onClick={() => {}}>
                 Extender sesión
               </Button>
-              <Button variant="outlined" color="error" onClick={handleLogOut}>
+              <Button variant="outlined" color="error" onClick={() => {}}>
                 Cerrar sesión
               </Button>
             </Box>
