+++ src/modules/shared/hooks/useSpecialtyQuery.ts
@@ -1,10 +1,10 @@
-import { specialtyService } from '@api/services';
 import { useQuery } from '@tanstack/react-query';
+import { getAllSpecialty } from '../services/specialty';
 
 const useGetSpecialtiesQuery = (enabled: boolean = true) => {
   return useQuery({
     queryKey: ['specialties'],
-    queryFn: () => specialtyService.getAllSpecialty(),
+    queryFn: () => getAllSpecialty(),
     enabled,
   });
 };
