+++ src/modules/shared/services/specialty/index.ts
@@ -1,5 +1,4 @@
-import specialtyEndpoints from '@api/endpoints/specialty';
-import { axiosInstance } from '@api/axios/axiosInstance';
+import { axiosInstance } from '@api/axiosInstance';
 import { handleError } from 'src/modules/shared/utils/handleError';
 
 import type { IBasicIdNameDescription } from 'src/modules/shared/types/common.types';
@@ -7,11 +6,10 @@ import type { AxiosError } from 'axios';
 
 export const getAllSpecialty = async () => {
   try {
-    const response = await axiosInstance.get<IBasicIdNameDescription[]>(
-      specialtyEndpoints.getSpecialties,
-    );
+    const response =
+      await axiosInstance.get<IBasicIdNameDescription[]>('/specialty');
     return response;
   } catch (error) {
-    return handleError(error as AxiosError);
+    return handleError(error as AxiosError<{ message?: string }>);
   }
 };
