--- src/page/Users/List/styles.css
@@ -1,14 +1,0 @@
-.header-users {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1.25rem;
-  margin-bottom: 1rem;
-}
-
-@media screen and (min-width: 768px) {
-  .header-users {
-    flex-direction: row;
-    justify-content: space-between;
-  }
-}
