--- src/page/auth/SignIn/styles.css
@@ -1,75 +1,0 @@
-.sign-in {
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.sign-in .container {
-  display: grid;
-  align-items: center;
-  justify-items: center;
-}
-
-.sign-in .form {
-  display: flex;
-  flex-direction: column;
-  border-radius: 0.4rem;
-}
-
-.sign-in .logo {
-  width: 4rem;
-  height: 4rem;
-  margin-bottom: 0.6rem;
-  align-self: self-start;
-}
-
-.sign-in .title {
-  margin-bottom: 0.1rem;
-}
-
-.sign-in .sub-title {
-  color: var(--clr-text-a50);
-  margin-bottom: 2.4rem;
-}
-
-.sign-in .inputs-group {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  width: 100%;
-}
-
-.sign-in .forget-password {
-  align-self: flex-end;
-  margin: 0.4rem 0 2rem;
-}
-
-.sign-in .buttons-group {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.sign-in .cover {
-  display: none;
-  height: 100vh;
-}
-
-@media screen and (min-width: 1024px) {
-  .sign-in .container {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 2rem;
-  }
-
-  .sign-in .cover {
-    display: flex;
-  }
-
-  .sign-in .cover-image {
-    width: 100%;
-    height: 100%;
-    background-size: cover;
-  }
-}
